+++ lace/include/lace/tree/SemanticAnalysis.h
@@ -65,7 +65,6 @@ public:
     void visit(CastExpr& node) override;
     void visit(ParenExpr& node) override;
     void visit(RefExpr& node) override;
-    void visit(SpecifierExpr& node) override;
     void visit(StructInitExpr& node) override;
     void visit(SubscriptExpr& node) override;
 };