+++ lace/include/lace/tree/SymbolAnalysis.h
@@ -32,8 +32,6 @@ public:
 
     void visit(SizeofExpr& node) override;
 
-    void visit(SpecifierExpr& node) override;
-
     void visit(StructInitExpr& node) override;
 };