+++ lace/source/codegen/LIRCodegenAddressed.cpp
@@ -25,8 +25,6 @@ lir::Value* LIRCodegen::codegen_addressed_expression(const Expr* expr) {
         return codegen_addressed_access(access);
     } else if (auto ref = dynamic_cast<const RefExpr*>(expr)) {
         return codegen_addressed_reference(ref);
-    } else if (auto spec = dynamic_cast<const SpecifierExpr*>(expr)) {
-        return codegen_addressed_specifier(spec);
     } else if (auto subscript = dynamic_cast<const SubscriptExpr*>(expr)) {
         return codegen_addressed_subscript(subscript);
     } else if (auto call = dynamic_cast<const CallExpr*>(expr)) {
@@ -88,12 +86,6 @@ lir::Value* LIRCodegen::codegen_addressed_reference(const RefExpr* expr) {
     }
 
     return nullptr;
-}
-
-lir::Value* LIRCodegen::codegen_addressed_specifier(const SpecifierExpr* expr) {
-    assert(expr->is_resolved());
-
-    return codegen_addressed_expression(expr->expr());
 }
 
 lir::Value* LIRCodegen::codegen_addressed_subscript(const SubscriptExpr* expr) {