+++ lace/source/lace.cpp
@@ -333,8 +333,8 @@ int32_t main(int32_t argc, char* argv[]) {
     log::direct(std::cout);
 
     std::vector<InputFile> files = {
-        InputFile("/home/lovelace/stl/test.lace"),
-        InputFile("/home/lovelace/stl/math.lace"),
+        InputFile("/home/lovelace/stl/linux.lace"),
+        InputFile("/home/lovelace/stl/mem.lace"),
     };
 
     for (int32_t i = 1; i < argc; ++i) {